+++ app/decorators/contact_decorator.rb
@@ -1,4 +1,4 @@
-class ContactsDecorator < Draper::Decorator
+class ContactDecorator < Draper::Decorator
   delegate_all
 
   def first_name
@@ -14,5 +14,15 @@ class ContactsDecorator < Draper::Decorator
     vector = vectors.email.first unless vector
     vector = vectors.first unless vector
     vector && vector.value
+  end
+
+  def status
+    return 'added' if added?
+    return 'rejected' if rejected?
+    ''
+  end
+
+  def notification_status
+    object.notifications.first.try(:status) || ''
   end
 end