+++ app/models/contact/extensions/filters.rb
@@ -1,4 +1,4 @@
-module Contact::FilterExtension
+module Contact::Extensions::Filters
   def notified?
     Notification.match_by_contact?(self)
   end