+++ app/models/contact/extensions/owners.rb
@@ -1,4 +1,4 @@
-module Contact::OwnerExtension
+module Contact::Extensions::Owners
   def owner
     Owner.new(owner_mkey)
   end