+++ spec/controllers/web_client_controller_spec.rb
@@ -1,9 +1,10 @@
 require 'rails_helper'
 
 RSpec.describe WebClientController, type: :controller, authenticate_with_http_basic: true do
-  let(:contact) { FactoryGirl.create :contact }
-  let(:notification) { FactoryGirl.create :notification, contact: contact }
+  let(:contact) { FactoryGirl.create(:contact) }
+  let(:notification) { FactoryGirl.create(:notification, contact: contact) }
   let(:nkey) { notification.nkey }
+  let(:owner) { Owner.new(contact.owner_mkey) }
 
   render_views
 
@@ -11,7 +12,7 @@ RSpec.describe WebClientController, type: :controller, authenticate_with_http_ba
     context 'with correct nkey' do
       before { get :show, id: nkey }
 
-      it_behaves_like 'response status'
+      it { expect(response).to have_http_status 200 }
       it { expect(response).to render_template(:show) }
     end
 
@@ -23,32 +24,29 @@ RSpec.describe WebClientController, type: :controller, authenticate_with_http_ba
   describe 'GET #add' do
     before { get :add, id: nkey }
 
-    it_behaves_like 'response status'
-    it { expect(response).to render_template(:action_handeled) }
+    it_behaves_like 'response redirect'
     it { expect(notification.reload.status).to eq 'added' }
   end
 
   describe 'GET #ignore' do
     before { get :ignore, id: nkey }
 
-    it_behaves_like 'response status'
-    it { expect(response).to render_template(:action_handeled) }
+    it_behaves_like 'response redirect'
     it { expect(notification.reload.status).to eq 'ignored' }
   end
 
   describe 'GET #unsubscribe' do
     before { get :unsubscribe, id: nkey }
 
-    it_behaves_like 'response status'
-    it { expect(response).to render_template(:action_handeled) }
-    it { expect(notification.reload.status).to eq 'unsubscribed' }
+    it_behaves_like 'response redirect'
+    it { expect(owner.unsubscribed?).to be true }
   end
 
   describe 'GET #subscribe' do
-    let!(:notification) { FactoryGirl.create :notification, contact: contact, status: 'unsubscribed' }
+    let!(:notification) { FactoryGirl.create(:notification, contact: contact) }
     before { get :subscribe, id: notification.nkey }
 
-    it { expect(notification.reload.status).to eq 'ignored' }
-    it { expect(response).to redirect_to(web_client_path) }
+    it_behaves_like 'response redirect'
+    it { expect(owner.subscribed?).to be true }
   end
 end