+++ spec/models/owner_spec.rb
@@ -1,10 +1,10 @@
 require 'rails_helper'
 
 RSpec.describe Owner, type: :model do
+  let!(:contact_1) { FactoryGirl.create(:contact, owner_mkey: mkey, total_score: 4) }
+  let!(:contact_2) { FactoryGirl.create(:contact, owner_mkey: mkey, total_score: 6) }
   let(:mkey) { 'xxxxxxxxxxxx' }
-  let!(:contact_1) { FactoryGirl.create :contact, owner_mkey: mkey, total_score: 4 }
-  let!(:contact_2) { FactoryGirl.create :contact, owner_mkey: mkey, total_score: 6 }
-  let(:instance) { described_class.new mkey }
+  let(:instance) { described_class.new(mkey) }
 
   describe '#full_name' do
     let(:mkey) { '7qdanSEmctZ2jPnYA0a1' }
@@ -29,27 +29,5 @@ RSpec.describe Owner, type: :model do
     end
 
     it { is_expected.to eq [contact_1] }
-  end
-
-  describe '#unsubscribed?' do
-    subject { instance.unsubscribed? }
-
-    context 'when unsubscribed' do
-      before do
-        FactoryGirl.create(:notification, contact: contact_1, status: 'added')
-        FactoryGirl.create(:notification, contact: contact_2, status: 'unsubscribed')
-      end
-
-      it { is_expected.to be true }
-    end
-
-    context 'when not unsubscribed' do
-      before do
-        FactoryGirl.create(:notification, contact: contact_1, status: 'ignored')
-        FactoryGirl.create(:notification, contact: contact_2, status: 'added')
-      end
-
-      it { is_expected.to be false }
-    end
   end
 end