+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+### v0.5.0
+- :bulb: Added status field for contacts concepts
+- :bulb: Implemented filter system for contacts concepts
+- :bulb: Implemented feature to ignore any contact on web-client
+- :arrow_up: Updated web-client navbar
+- :arrow_up: Updated web-client notification messages
+- :arrow_up: Refactored whole web-client
+- :arrow_up: Created model for storing owner data (unsubscribed status)
+- :hammer: Fixed dockerfile assets precompile
+- :hammer: Fixed minor bugs
+
 ### v0.4.1
 - :bulb: Added feature to update all contacts by owner
 - :hammer: Fixed images assets